requirements.txt
appdirs==1.4.4
certifi==2024.8.30
charset-normalizer==3.4.0
future==1.0.0
geographiclib==2.0
gpxpy==1.6.2
idna==3.10
lxml==5.3.0
pillow==11.0.0
py-staticmaps==0.4.0
pycairo==1.27.0
python-docx==1.1.2
python-slugify==8.0.4
requests==2.32.3
s2sphere==0.2.5
svgwrite==1.4.3
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
